Alterando models e criando repository
