+++ app/controllers/login_controller.py
@@ -1,7 +1,7 @@
 from flask import redirect, flash, session
 from flask_bcrypt import check_password_hash
 
-from app.models.user import User
+from app.repository.user_repository import UserRepository
 
 
 class LoginController:
@@ -10,17 +10,18 @@ class LoginController:
         return redirect("/")
 
     def authenticate(self, view, request):
-        email = request.form['email']
-        senha = request.form['senha']
+        try:
+            email = request.form['email']
+            senha = request.form['senha']
 
-        user = User.check_if_email_exists(email)
-        password = check_password_hash(user.senha, senha)
+            user = UserRepository.check_if_email_exists(email)
+            password = check_password_hash(user.senha, senha)
 
-        if user and password:
-                session['usuario_logado'] = user.id
-                flash(f"{user.nome} logou com sucesso!")
-                return redirect("/importar-transacoes")
-        else:
+            if user and password:
+                    session['usuario_logado'] = user.id
+                    flash(f"{user.nome} logou com sucesso!")
+                    return redirect("/importar-transacoes")
+        except:
             flash("Usuário não logado, tente novamente!")
             return redirect('/')
